requirements.txt
web3==6.11.1
python-dotenv==1.0.0
typing-extensions==4.7.1
requests==2.31.0 